backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
anthropic==0.18.1
python-dotenv==1.0.0
schwab-py==1.5.0
httpx==0.25.2
langgraph==0.0.55
pyfinviz==0.1.4
pandas==2.1.4
praw==7.7.1